WIP: convert JS files to ES6
Had to remove all the sprockets includes from the tests to stop rails
from panicking about the lack of JS files in the app/assets.

Installed the npm version of frontend toolkit

Had to use a committed version of trap because it's not available in
jspm / npm.

TODO: analytics
TODO: get the tests running in node
TODO: beautify the code
TODO: production build

Solo: @richardtowers