5032: Store idps in session after authn_request
This will allow us to avoid an API call on the about-certified-companies
and sign-in pages.

This must be released after the API is updated to have idps on POST
/api/session.

Once this is released and all the sessions have rolled through we can
remove calls to /api/session/{id}/federation from new frontend.

Authors: @oswaldquek @richardtowers